fix copy paste errors
